+++ kikuchipy/signals/tests/test_ebsd.py
@@ -1472,7 +1472,7 @@ class TestEBSDRefinement:
         )
         detector = kp.detectors.EBSDDetector(shape=s.axes_manager.signal_shape[::-1])
         xmap.phases[0].name = self.mp.phase.name
-        xmap_refined = s.refine_orientation(
+        dask_array = s.refine_orientation(
             xmap=xmap,
             master_pattern=self.mp,
             energy=20,
@@ -1480,9 +1480,10 @@ class TestEBSDRefinement:
             method_kwargs=dict(options=dict(maxiter=10)),
             compute=False,
         )
-        assert isinstance(xmap_refined, list)
-        assert dask.is_dask_collection(xmap_refined[0])
-        assert len(xmap_refined) == 9
+        assert isinstance(dask_array, da.Array)
+        assert dask.is_dask_collection(dask_array)
+        # Should ideally be (3, 3, 4) with better use of map_blocks()
+        assert dask_array.shape == s.axes_manager.navigation_shape[::-1] + (1,)
 
     @pytest.mark.filterwarnings("ignore: The line search algorithm did not converge")
     @pytest.mark.filterwarnings("ignore: Angles are assumed to be in radians, ")
@@ -1583,6 +1584,69 @@ class TestEBSDRefinement:
             xmap_refined_no_mask.rotations.data, xmap_refined_mask.rotations.data
         )
 
+    @pytest.mark.parametrize(
+        "ebsd_with_axes_and_random_data, detector, rechunk, chunk_kwargs, chunksize",
+        [
+            (
+                ((5, 4), (10, 8), True, np.float32),
+                ((5, 4), (10, 8)),
+                False,
+                None,
+                (5, 4, 1),
+            ),
+            (
+                ((5, 4), (10, 8), True, np.float32),
+                ((5, 4), (10, 8)),
+                True,
+                dict(chunk_shape=3),
+                (3, 3, 1),
+            ),
+            (
+                ((5, 4), (10, 8), True, np.float32),
+                ((5, 4), (10, 8)),
+                False,
+                dict(chunk_shape=3),
+                (5, 4, 1),
+            ),
+        ],
+        indirect=["ebsd_with_axes_and_random_data", "detector"],
+    )
+    def test_refine_orientation_chunking(
+        self,
+        ebsd_with_axes_and_random_data,
+        detector,
+        rechunk,
+        chunk_kwargs,
+        chunksize,
+        get_single_phase_xmap,
+    ):
+        """Ensure the returned dask array when not computing has the
+        desired chunksize.
+
+        Ideally, the last dimension should have size 4 (score, phi1,
+        Phi, phi2), but this requires better handling of removed and
+        added axes and their sizes in the call to
+        :func:`dask.array.map_blocks` in :func:`_refine_orientation` and
+        the other equivalent private refinement functions.
+        """
+        s = ebsd_with_axes_and_random_data
+        xmap = get_single_phase_xmap(
+            nav_shape=s.axes_manager.navigation_shape[::-1],
+            rotations_per_point=1,
+            step_sizes=tuple(a.scale for a in s.axes_manager.navigation_axes)[::-1],
+        )
+        xmap.phases[0].name = self.mp.phase.name
+        dask_array = s.refine_orientation(
+            xmap=xmap,
+            master_pattern=self.mp,
+            energy=20,
+            detector=detector,
+            compute=False,
+            rechunk=rechunk,
+            chunk_kwargs=chunk_kwargs,
+        )
+        assert dask_array.chunksize == chunksize
+
     # ------------------- Refine projection centers ------------------ #
 
     @pytest.mark.parametrize(
@@ -1618,7 +1682,6 @@ class TestEBSDRefinement:
             rotations_per_point=1,
             step_sizes=tuple(a.scale for a in s.axes_manager.navigation_axes)[::-1],
         )
-        print(xmap.scores.shape)
         xmap.phases[0].name = self.mp.phase.name
         method_kwargs.update(dict(options=dict(maxiter=10)))
         new_scores, new_detector = s.refine_projection_center(
@@ -1697,7 +1760,7 @@ class TestEBSDRefinement:
         )
         detector = kp.detectors.EBSDDetector(shape=s.axes_manager.signal_shape[::-1])
         xmap.phases[0].name = self.mp.phase.name
-        delayed_results = s.refine_projection_center(
+        dask_array = s.refine_projection_center(
             xmap=xmap,
             master_pattern=self.mp,
             energy=20,
@@ -1705,9 +1768,10 @@ class TestEBSDRefinement:
             method_kwargs=dict(options=dict(maxiter=10)),
             compute=False,
         )
-        assert isinstance(delayed_results, list)
-        assert dask.is_dask_collection(delayed_results[0])
-        assert len(delayed_results) == 9
+        assert isinstance(dask_array, da.Array)
+        assert dask.is_dask_collection(dask_array)
+        # Should ideally be (3, 3, 4) with better use of map_blocks()
+        assert dask_array.shape == (3, 3, 1)
 
     # ---------- Refine orientations and projection centers ---------- #
 
@@ -1811,7 +1875,7 @@ class TestEBSDRefinement:
         )
         xmap.phases[0].name = self.mp.phase.name
         detector = kp.detectors.EBSDDetector(shape=s.axes_manager.signal_shape[::-1])
-        delayed_results = s.refine_orientation_projection_center(
+        dask_array = s.refine_orientation_projection_center(
             xmap=xmap,
             master_pattern=self.mp,
             energy=20,
@@ -1819,9 +1883,10 @@ class TestEBSDRefinement:
             method_kwargs=dict(options=dict(maxiter=1)),
             compute=False,
         )
-        assert isinstance(delayed_results, list)
-        assert dask.is_dask_collection(delayed_results[0])
-        assert len(delayed_results) == 9
+        assert isinstance(dask_array, da.Array)
+        assert dask.is_dask_collection(dask_array)
+        # Should ideally be (3, 3, 7) with better use of map_blocks()
+        assert dask_array.shape == (3, 3, 1)
 
 
 class TestAverageNeighbourDotProductMap: